+++ babble-react/public/index.html
@@ -30,7 +30,7 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="babble" data-username="jeeku" data-valid="true"></div>
+    <div id="babble" data-username="jeeku" data-valid="false"></div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
